[VSPC-99] PR review comments
